Fixing up PEAP authentications
This also refactors SSL functions into a common file tls.c/.h and re-uses them
in both peap and eap files. Functions that are common are the initialization
functions such as set CA verify location, CRL files, SSL options, TLS method,
and the default verify callback.

This is to make the options beteen eap-tls and peap common and to adher to a
common set of SSL options provided to PPP for all of CA, CRL, Verify, etc.

Signed-off-by: Eivind Næss <eivnaes@yahoo.com>