pppoe: Remove rp-pppoe.so symlink to not conflict with real rp-pppoe.so plugin (#304)
Backward compatibility symlink is there already for one ppp release. Remove
it for next ppp release to not conflict with real rp-pppoe.so plugin. So
both ppp's pppoe.so and rp's rp-pppoe.so plugins can be installed together.

Now when conversion to automake was done, it is a good time to drop this
problematic symlink from default installation.

Signed-off-by: Pali Rohár <pali@kernel.org>
Signed-off-by: Adrian Ban <devel@easynet.dev>