Let ./configure control the paths for pppd
This was previously done by specifying an overriding value for _ROOT_PATH. With this change, this variable is now gone.
Instead, pathnames.h will use the SYSCONFDIR and LOCALSTATEDIR to resolve these paths. These directories is already controlled by
configure.

Package maintainers should be aware though that this may change their current configuration. The convential ./configure way is to
specify:
   ./configure --prefix=/usr --sysconfdir=/etc --localstatedir=/var/run --with-plugin-dir=/usr/lib/pppd/2.4.10

If one omit the --sysconfdir option, then the default location is by ${prefix}/etc which may not be what you want.

Signed-off-by: Eivind Næss <eivnaes@yahoo.com>