pppd: add noresolvconf option
On embedded systems, it is easier to avoid the resolv.conf conflicts
than to fix them.
'noresolvconf' option on top of 'usepeerdns' lets pppd request DNS information
and forward the info to the scripts, but will not write resolv.conf.
The ip-up/ip-down scripts can then deal with the info and no conflict ever
existed.

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>