plugins/pppoe: Remove 5 second delay before PADR in verbose mode
This partially reverts commit 1c082acf77e6 ("pppoe: Show verbose
information about all concentrator when pppoe-verbose option is set",
2021-01-01).  That commit added a 5-second wait between receiving a
valid PADO and sending the PADR response so as to give time to see
whether any other PADOs arrive.  However, it appears that this delay
causes problems with some concentrators.

This reverts to the previous behaviour of sending the PADR
immediately.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>