plugins/pppoe: Export AC name to scripts via an environment variable
This saves the access concentrator (AC) name supplied by the AC in the
PADO packet and creates an environment variable called "ACNAME" with
the name as its value for scripts to use if desired.

This was inspired by a pull request from "bearmi" on github, but
reimplemented somewhat differently by me.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>