Remove Solaris module code
This code has not been updated for a long time, and nominally targets
an OS which Oracle (its owner) has all but abandoned.  It includes
zlib code which has CVEs against it, and it is unknown whether any
of them would be exploitable in this context.

Illumos, which is an OpenSolaris fork, has forked Solaris-native
drivers, which should probably work with the user-space pppd code
here.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>