chat: Add copyright notice
This adds a copyright notice in my name with MIT licence permissions
to this file.  The reason is primarily to provide a concrete statement
of permission to use, copy, modify and distribute the program, for the
sake of those who are uncertain about meaning of the public domain.
See for example https://github.com/ppp-project/ppp/issues/474 .
Given that I have maintained this program for 25 years, it seems
reasonable for it to be my copyright on it.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>