pppd: add VRF (Virtual Routing and Forwarding) binding support on Linux
Add 'vrf' option to bind a PPP interface to a specific VRF, so that
routes are installed in the VRF's routing table rather than the main
routing table.

This allows PPP sessions to operate in separate routing domains,
providing isolation of traffic and routes.

When this option is used, pppd sets the 'VRF' environment variable for
ip-{up,down} scripts.

Signed-off-by: Martin Olivier <martin.olivier@live.fr>