dhcpv6relay: When responding to router solicitations, do so using unicast.
https://datatracker.ietf.org/doc/html/rfc4861#section-4.2 states:

Destination Address
    Typically the Source Address of an invoking Router
    Solicitation or the all-nodes multicast address.

This implies we should respond *to* the requesting device only using
unicast, thus make it so.

Use recvfrom to achieve this, also making the socket non-blocking.

If we fail to receive due to EWOULDBLOCK or EAGAIN - assume we "lost" a
RS and just send a RA to multicast anyway.

Signed-off-by: Jaco Kroon <jaco@uls.co.za>