dhcpv6relay: Improve buffer overrun protection by using MSG_TRUNC.
When calling recvfrom we can pass MSG_TRUNC to always get the number of
bytes that was in the original packet (for non-connection oriented
protocols such as UDP) rather than the number of bytes placed into the
buffer.  This allows us to compare the number of bytes in the packet to
the size of the buffer rather than just the number of bytes placed in
the buffer, allowing us to receive and process a full buffer rather than
just 1 byte less than the buffer size.

Technically this fixes an off by one error, but there should be no
operational impact and highly unlikely to occur in the wild.

Signed-off-by: Jaco Kroon <jaco@uls.co.za>